+++ app/controllers/services/questions_controller.rb
@@ -4,7 +4,8 @@ class Services::QuestionsController < ApplicationController
   def create
     @service = Service.friendly.find(params[:service_id])
     @service.contact_emails.each  do |email|
-      ServiceMailer.new_question(email, params[:service_question], @service).deliver_now
+      ServiceMailer.new_question(email, current_user,
+                                 params[:service_question], @service).deliver_now
     end
 
     redirect_to service_path(@service)