+++ spec/features/backoffice/services_spec.rb
@@ -42,6 +42,8 @@ RSpec.feature "Services in backoffice" do
   scenario "I can create new service", js: true do
     provider = create(:provider)
     research_area = create(:research_area)
+    platform = create(:platform)
+
     visit backoffice_services_path
     click_on "Create new service"
 
@@ -65,8 +67,8 @@ RSpec.feature "Services in backoffice" do
     fill_in "Activate message", with: "Welcome!!!"
     select research_area.name, from: "Research areas"
     select provider.name, from: "Providers"
-
-    check "Open access"
+    select "open_access", from: "Service type"
+    select platform.name, from: "Platforms"
     fill_in "service_contact_emails_0", with: "person1@test.ok"
     page.find("#add-email-field").click
     fill_in "service_contact_emails_1", with: "person2@test.ok"
@@ -79,7 +81,7 @@ RSpec.feature "Services in backoffice" do
     expect(page).to have_content("service terms of use")
     expect(page).to have_content("service tagline")
     expect(page).to have_content("https://sample.url")
-    expect(page).to have_content("true")
+    expect(page).to have_content("open_access")
     expect(page).to have_content("person1@test.ok")
     expect(page).to have_content("person2@test.ok")
     expect(page).to have_content("Welcome!!!")