+++ spec/models/platform_spec.rb
@@ -1,0 +1,7 @@
+# frozen_string_literal: true
+
+require "rails_helper"
+
+RSpec.describe Platform, type: :model do
+  it { should validate_presence_of(:name) }
+end